// src/models/enums.ts
export enum InputType {
  button = "button",
  checkbox = "checkbox",
  color = "color",
  date = "date",
  datetimelocal = "datetime-local",
  email = "email",
  file = "file",
  hidden = "hidden",
  image = "image",
  month = "month",
  number = "number",
  password = "password",
  radio = "radio",
  range = "range",
  reset = "reset",
  search = "search",
  submit = "submit",
  tel = "tel",
  text = "text",
  time = "time",
  url = "url",
  week = "week",
  select = "select",
}

export enum CountryEnum {
  USA = "United States of America",
  EUROPE = "Europe",
  JAPAN = "Japan",
  UK = "United Kingdom",
  LATIN = "Latin America",
}

export enum AdminEnum {
  addCasestudy = "addCasestudy",
  updateCasestudy = "updateCasestudy",
  deleteCasestudy = "deleteCasestudy",
  addCasestudyContent = "addCasestudyContent",
  updateCasestudyContent = "updateCasestudyContent",
  deleteCasestudyContent = "deleteCasestudyContent",
  addCasestudyGallery = "addCasestudyGallery",
  updateCasestudyGallery = "updateCasestudyGallery",
  deleteCasestudyGallery = "deleteCasestudyGallery",
}